Updates Added

+++ STM32M4/src/HAL/include/SWITCH/SWITCH.h
@@ -11,8 +11,8 @@
 
 
 typedef unsigned char u8;
-typedef unsigned int u16;
-typedef unsigned long u32;
+typedef unsigned short u16;
+typedef unsigned int u32;
 
 typedef struct 
 {
