+++ STM32M4/src/MCAL/include/NVIC.h
@@ -49,8 +49,8 @@
 
 
 typedef unsigned char u8;
-typedef unsigned int u16;
-typedef unsigned long u32;
+typedef unsigned short u16;
+typedef unsigned int u32;
 
 /*NVIC_Registers*/
